+++ Src/App/Commands/AccountRemoveCommand.cs
@@ -1,0 +1,45 @@
+﻿using BMBank.Src.App.Commands.Interfaces;
+using BMBank.Src.DatabaseInteraction.Core.DAO;
+using BMBank.Src.Network;
+using BookOrg.Src.Logic.Connection;
+
+namespace BMBank.Src.App.Commands
+{
+    public class AccountRemoveCommand : ICommand, IAccountInteractionCommand
+    {
+        public string Key => "AR";
+        public AccountDAO DAO { get; }
+
+        public AccountRemoveCommand(AccountDAO dao)
+        {
+            DAO = dao;
+        }
+
+
+        public string Execute(string arguments)
+        {
+            string[] accountParts = arguments.Trim().Split('/');
+            if (accountParts.Length != 2)
+            {
+                throw new FormatException("Invalid format. Usage: AR <account>/<ip>");
+            }
+
+            if (!int.TryParse(accountParts[0], out int accountNumber))
+            {
+                throw new FormatException("Invalid account number.");
+            }
+            string targetIp = accountParts[1];
+
+            string? localIp = IPAddressObtainer.GetLocalIPv4Address();
+
+            if (targetIp != localIp)
+            {
+                throw new InvalidOperationException("Invalid ip address provided.");
+            }
+
+            DAO.RemoveAccount(accountNumber);
+
+            return "AR";
+        }
+    }
+}