+++ Src/App/Commands/AccountWithdrawCommand.cs
@@ -1,53 +1,55 @@
 ﻿using BMBank.Src.App.Commands.Interfaces;
 using BMBank.Src.DatabaseInteraction.Core.DAO;
 using BMBank.Src.Network;
+
 namespace BMBank.Src.App.Commands
 {
-    public class AccountDepositCommand: ICommand, IAccountInteractionCommand
+    public class AccountWithdrawCommand : ICommand, IAccountInteractionCommand
     {
-        public string Key => "AD";
+        public string Key => "AW";
         public AccountDAO DAO { get; }
 
-        public AccountDepositCommand(AccountDAO dao)
+
+        public AccountWithdrawCommand(AccountDAO dao)
         {
             DAO = dao;
         }
 
+
         public string Execute(string arguments)
         {
-            string[] argumentParts = arguments.Split(' ', StringSplitOptions.RemoveEmptyEntries);
-            if (argumentParts.Length != 2)
+            string[] argumentsParts = arguments.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+            if (argumentsParts.Length != 2)
             {
-                throw new FormatException("Invalid format. Usage: AD <account>/<ip> <amount>");
+                throw new FormatException("Invalid format. Usage: AW <account>/<ip> <amount>");
             }
 
-            string[] accountParts = argumentParts[0].Split('/');
-
+            string[] accountParts = argumentsParts[0].Split('/');
             if (accountParts.Length != 2)
-            { 
+            {
                 throw new FormatException("Invalid account format.");
             }
 
-            if (!int.TryParse(accountParts[0], out int accountNumber)) 
+            if (!int.TryParse(accountParts[0], out int accountNumber))
             {
                 throw new FormatException("Invalid account number.");
             }
             string targetIp = accountParts[1];
 
-            if (!long.TryParse(argumentParts[1], out long amount))
-            { 
+            if (!long.TryParse(argumentsParts[1], out long amount))
+            {
                 throw new FormatException("Invalid amount.");
             }
 
             string? localIp = IPAddressObtainer.GetLocalIPv4Address();
             if (targetIp != localIp)
-            { 
+            {
                 throw new InvalidOperationException("Invalid ip address provided.");
             }
 
-            DAO.Deposit(accountNumber, amount);
+            DAO.Withdraw(accountNumber, amount);
 
-            return "AD";
+            return "AW";
         }
     }
 }